backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg[binary]==3.2.10
SQLAlchemy==2.0.35
Werkzeug==3.0.4
Flask-CORS==4.0.1
python-dotenv==1.0.1